+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+
+## 1.1.0 - 
+* generateStateTransition renamed to generateStateDiagramTransition (minor breaking change)
+* assertCannotReach made static and now accessible via FsmRx derived classes (minor breaking change)
+* destroy$, nextChangeStateTransition$, override$ now return as observables from a getter rather than subjects.(minor breaking change) 
+* resolvedFsmConfig and isInDevMode made readonly. 
+* nextChangeStateTransition$ now omits the data of the state being left. 
+* added currentStatePostInit$ getter which pre-filters the init state. 
+* added FsmRxConcrete for use with functional components. 
+
 ## 1.0.1 - 8/4/2024
 * Added CHANGELOG.md
 * Updated README.md file